@@ -10,44 +10,39 @@ long_description_content_type = text/markdown
 license = MIT
 license_file = LICENSE
 classifiers =
-  Development Status :: 2 - Pre-Alpha
+  Development Status :: 3 - Alpha
   Environment :: Console
   Intended Audience :: Science/Research
   License :: OSI Approved :: MIT License
   Operating System :: POSIX
   Programming Language :: Python
-  Programming Language :: Python :: 2
-  Programming Language :: Python :: 2.7
   Programming Language :: Python :: 3
-  Programming Language :: Python :: 3.5
   Programming Language :: Python :: 3.6
   Programming Language :: Python :: 3.7
   Programming Language :: Python :: 3.8
   Programming Language :: Python :: 3.9
+  Programming Language :: Python :: 3.10
+  Programming Language :: Python :: 3.11
   Topic :: Internet :: WWW/HTTP
 
 [options]
 zip_safe = false
 packages = find:
 include_package_data=true
-python_requires = >=2.7
+python_requires = >=3.6
 setup_requires =
   setuptools >= 30.3.0
   pytest-runner
 install_requires =
-  future; python_version<="2.7"
-  six >= 1.10
   setuptools >= 21.0.0
   wheel >= 0.37.0
-  requests >= 2.28.0
-  requests-toolbelt >= 0.9.0
   click >= 3.0
   PyYAML >= 5.1
-  gql >= 2.0.0, < 3 # Dependency is broken in version 3.0.0
+  gql[all] >= 3.0.0
 tests_require =
-  pytest~=4.6.7 # needed for python 2.7+3.5
+  pytest>=4.6.7
   pytest-cov>=2.8.1
-  pytest-randomly==1.2.3 # needed for python 2.7+3.5
+  pytest-randomly>=1.2.3
 
 [options.extras_require]
 develop =
